requirements.txt
nsepython==1.0.6
schedule==1.2.0
pandas==2.2.2
numpy==1.26.4
python-telegram-bot==13.15
